+++ interval_ccd/interval.hpp
@@ -39,7 +39,7 @@ typedef boost::numeric::interval<
 #endif
 
 typedef std::array<std::pair<int,int>,3> Paraccd;//<k,n> pair present parameters (u,v,t) which are presented as k/(2^n) 
-typedef std::pair<int,int> Numccd;//<k,n> pair present a number k/pow(2,n)
+typedef std::pair<long,int> Numccd;//<k,n> pair present a number k/pow(2,n)
 typedef std::pair<Numccd,Numccd> Singleinterval;// a interval presented by two double numbers
 typedef std::array<Singleinterval,3> Interval3; // 3 dimesional interval
 typedef Eigen::Matrix<Rational, 3, 1, Eigen::ColMajor | Eigen::DontAlign> Vector3r;
